+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.8 [????.??.??]
+* Require `singletons-base-3.0` and GHC 9.0.
+* Remove eliminators for `Data.Semigroup.Option`, which is deprecated as of
+  `base-4.15.0.0`.
+
 ## 0.7 [2020.03.25]
 * Require `singletons-2.7` and GHC 8.10.
 * Add experimental support for generating type-level eliminators through the